+++ lib/alg/gltf.types.ts
