--- lib/wrappers/glb-renderer.types.ts
@@ -1,12 +1,0 @@
-export interface GlbRendererRawData {
-    readonly data: ArrayBufferView;
-}
-
-export interface GlbRendererUrlData {
-    readonly url: string;
-}
-
-export type GlbRendererData = (
-    | GlbRendererRawData
-    | GlbRendererUrlData
-);
