+++ lib/wrappers/gltf-renderer.ts
@@ -1,14 +1,14 @@
-import type { GlbRendererData, GlbRendererRawData, GlbRendererUrlData } from './glb-renderer.types';
+import type { GlTFRendererData, GlTFRendererRawData, GlTFRendererUrlData } from './gltf-renderer.types';
 import type { Runtime } from '../gl/runtime';
 import { BaseWrapper } from '../gl/base-wrapper';
-import { parseGlTF } from '../alg/glb';
+import { parseGlTF } from '../alg/gltf';
 
-function isRawData(data: GlbRendererData): data is GlbRendererRawData {
-    return data && ArrayBuffer.isView((data as GlbRendererRawData).data);
+function isRawData(data: GlTFRendererData): data is GlTFRendererRawData {
+    return data && ArrayBuffer.isView((data as GlTFRendererRawData).data);
 }
 
-function isUrlData(data: GlbRendererData): data is GlbRendererUrlData {
-    return data && typeof (data as GlbRendererUrlData).url === 'string';
+function isUrlData(data: GlTFRendererData): data is GlTFRendererUrlData {
+    return data && typeof (data as GlTFRendererUrlData).url === 'string';
 }
 
 export class GlbRenderer extends BaseWrapper {
@@ -16,7 +16,7 @@ export class GlbRenderer extends BaseWrapper {
         super(runtime.logger(), tag);
     }
 
-    async setData(data: GlbRendererData): Promise<void> {
+    async setData(data: GlTFRendererData): Promise<void> {
         if (!data) {
             throw this._logger.error('set_data: null');
         }
