+++ lib/wrappers/gltf-renderer.types.ts
@@ -1,0 +1,12 @@
+export interface GlTFRendererRawData {
+    readonly data: ArrayBufferView;
+}
+
+export interface GlTFRendererUrlData {
+    readonly url: string;
+}
+
+export type GlTFRendererData = (
+    | GlTFRendererRawData
+    | GlTFRendererUrlData
+);
